requirements.txt
grpcio==1.14.0
grpcio-tools==1.14.0
gevent==1.4.0